Make Tab Panels work
The all-in-one Teaching Services page can be accessed from several links in the main navigation and now it correctly renders the selected tab.  Adding the key={tabNumber} was the key to getting the ServiceTabs to rerender when selecting a new topic from the main nav while the teachPage is already mounted.